setup route

+++ app/controllers/PhotosController.js
@@ -2,6 +2,48 @@ const { PrismaClient } = require("@prisma/client");
 const fs = require("fs");
 
 const prisma = new PrismaClient();
+
+const getAllPhoto = async (req, res) => {
+  const data = await prisma.photos.findMany();
+  if (data.length === 0) {
+    return res.status(404).json({
+      status: 404,
+      success: false,
+      message: "Photos are empty",
+    });
+  }
+  return res.status(200).json({
+    status: 200,
+    success: true,
+    message: "Photos retrieved successfully",
+    data,
+  });
+};
+
+const getSinglePhoto = async (req, res) => {
+  const { id } = req.params;
+  const data = await prisma.photos.findUnique({
+    where: {
+      id: Number(id),
+    },
+  });
+  if (data === null) {
+    return res.status(404).json({
+      status: 404,
+      success: false,
+      message: "Image not found",
+    });
+  }
+  if (data) {
+    return res.status(200).json({
+      status: 200,
+      success: true,
+      message: "Image retrieved successfully",
+      data,
+    });
+  }
+};
+
 const uploadPhoto = async (req, res) => {
   if (req.file) {
     await prisma.photos.create({
@@ -13,14 +55,14 @@ const uploadPhoto = async (req, res) => {
     return res.status(200).json({
       status: 200,
       success: true,
-      message: "upload success",
+      message: "Upload success",
       data: req.file,
     });
   }
   return res.status(404).json({
     status: 404,
     success: false,
-    message: "no file",
+    message: "No file",
   });
 };
 
@@ -37,7 +79,7 @@ const uploadPhotos = async (req, res) => {
     return res.status(200).json({
       status: 200,
       success: true,
-      message: "upload success",
+      message: "Upload success",
       data: req.files,
     });
   }
@@ -55,21 +97,32 @@ const deletePhoto = async (req, res) => {
       id: Number(id),
     },
   });
-  fs.unlink(photo.path, async function () {
-    await prisma.photos.delete({
-      where: {
-        id: Number(id),
-      },
+  if (photo === null) {
+    return res.status(404).json({
+      status: 404,
+      success: false,
+      message: "Image not found",
     });
-    return res.status(200).json({
-      status: 200,
-      success: true,
-      message: "Delete Successfully",
+  }
+  if (photo) {
+    fs.unlink(photo.path, async function () {
+      await prisma.photos.delete({
+        where: {
+          id: Number(id),
+        },
+      });
+      return res.status(200).json({
+        status: 200,
+        success: true,
+        message: "Delete successfully",
+      });
     });
-  });
+  }
 };
 
 module.exports = {
+  getAllPhoto,
+  getSinglePhoto,
   uploadPhoto,
   uploadPhotos,
   deletePhoto,
